+++ src/main/kotlin/com/meetfootball/adapter/base/in/web/response/ApiResponse.kt
@@ -1,0 +1,3 @@
+package com.meetfootball.adapter.base.`in`.web.response
+
+abstract class ApiResponse
