+++ src/main/kotlin/com/meetfootball/application/user/domain/entity/UserEntity.kt
@@ -1,0 +1,7 @@
+package com.meetfootball.application.user.domain.entity
+
+class UserEntity(
+    val name: String,
+    val email: String,
+    val password: String
+)
