+++ src/main/kotlin/com/meetfootball/application/user/port/in/FindUserByIdUseCase.kt
@@ -1,0 +1,8 @@
+package com.meetfootball.application.user.port.`in`
+
+import com.meetfootball.application.base.Result
+import com.meetfootball.application.user.domain.entity.UserEntity
+
+interface FindUserByIdUseCase {
+    fun findUserById(id: Long): Result<UserEntity>
+}
