--- src/main/kotlin/com/meetfootball/common/adapter/Persistence.kt
@@ -1,13 +1,0 @@
-package com.meetfootball.common.adapter
-
-import org.springframework.core.annotation.AliasFor
-import org.springframework.stereotype.Component
-
-@Target(AnnotationTarget.CLASS)
-@Retention(AnnotationRetention.RUNTIME)
-@MustBeDocumented
-@Component
-annotation class Persistence(
-    @get:AliasFor(annotation = Component::class)
-    val value: String = "",
-)
