--- src/main/kotlin/com/meetfootball/common/enums/ErrorCode.kt
@@ -1,5 +1,0 @@
-package com.meetfootball.common.enums
-
-enum class ErrorCode(val code: String) {
-    NOT_FOUND("0404")
-}
