--- src/main/kotlin/com/meetfootball/common/jpa/JpaAuditing.kt
@@ -1,9 +1,0 @@
-package com.meetfootball.common.jpa
-
-import org.springframework.context.annotation.Configuration
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing
-
-@Configuration
-@EnableJpaAuditing
-class JpaAuditing {
-}
