--- src/main/kotlin/com/meetfootball/common/response/ApiResponse.kt
@@ -1,5 +1,0 @@
-package com.meetfootball.common.response
-
-import com.meetfootball.user.application.domain.entity.ResultEntity
-
-abstract class ApiResponse
