--- src/main/kotlin/com/meetfootball/common/response/ExceptionResponse.kt
@@ -1,9 +1,0 @@
-package com.meetfootball.common.response
-
-import com.meetfootball.user.application.domain.entity.ResultEntity
-
-data class ExceptionResponse<T>(
-    private val resultEntity: ResultEntity.Fail<T>,
-    val message: String = resultEntity.error.message,
-    val errorCode: String = resultEntity.error.errorCode.code,
-): ApiResponse()
