+++ src/main/kotlin/com/meetfootball/config/SwaggerConfig.kt
@@ -1,0 +1,20 @@
+package com.meetfootball.config
+
+import io.swagger.v3.oas.models.OpenAPI
+import io.swagger.v3.oas.models.info.Info
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+class SwaggerConfig {
+    @Bean
+    fun openAPI(): OpenAPI {
+        return OpenAPI()
+            .info(
+                Info()
+                    .title("Meet Football API")
+                    .version("1.0.0")
+                    .description("Meet Football 프로젝트 API Swagger")
+            )
+    }
+}
