+++ src/main/kotlin/com/meetfootball/enum/ErrorCode.kt
@@ -1,0 +1,5 @@
+package com.meetfootball.enum
+
+enum class ErrorCode(val code: String) {
+    NOT_FOUND("0404")
+}
