--- src/main/kotlin/com/meetfootball/user/adapter/in/web/response/FindUserByIdResponse.kt
@@ -1,13 +1,0 @@
-package com.meetfootball.user.adapter.`in`.web.response
-
-import com.meetfootball.common.response.ApiResponse
-import com.meetfootball.user.application.domain.entity.ResultEntity
-import com.meetfootball.user.application.domain.entity.User
-
-data class FindUserByIdResponse(
-    private val result: ResultEntity.Success<User>
-): ApiResponse() {
-    val name: String = result.data.name
-
-    val email: String = result.data.email
-}
