--- src/main/kotlin/com/meetfootball/user/application/domain/entity/User.kt
@@ -1,7 +1,0 @@
-package com.meetfootball.user.application.domain.entity
-
-class User(
-    val name: String,
-    val email: String,
-    val password: String
-)
